Update ThunderBasicAuto2016_2017LinearOpMode.java
This has the algorithm for the robot going to the tape and aligning itself to face the beacon. It doesn't have the rotation speed for the robot yet, nor does it have the things that compensate for the color sensor displacement. Will need to add those and the stuff for poking the beacon.